统计信息补充

+++ src/constant/meta.rs
@@ -23,4 +23,4 @@ pub const MARKDOWN_BASE_TITLE_TAG: &str = "### 母题";
 // 拆分母题变式题的标题等级
 pub const MARKDOWN_PREFIX_TITLE_LIST: [&str; 2] = ["### ", "#### "];
 // 统计数据最大数量
-pub const STAT_MAX_NUM: usize = 10;
+pub const STAT_MAX_NUM: i16 = 10;
